readme: List more important limitations
While the project seems very interesting at first, many people discover
it's not very helpful if it crashes on comments or empty lines.  Given
that the problems are known since 2015 (issues #1, #2), I think it's
only fair to list them in the 'limitations' section.  Arrays are also
an important missing feature (#24).